pyproject.toml
[project]
name = "live-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.63.0",
    "numpy>=2.4.2",
    "pyaudio>=0.2.14",
]
